abi-gen: mv Py Handlebars helpers to own module
Move all existing Python-related Handlebars helpers to the newly created
python_handlebars_helpers module.
